+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [06-16-2025]
+### Added
+- TodoLists basic look
+
+### Removed
+- Homepage
+
+
 ## [05-12-2025]
 ### Added
 - Snackbar for failed login